refactor:utils reconfiguration(#22)
