+++ htmlparser/src/helpers/attributeHelpers.ts
