+++ htmlparser/src/helpers/stringHelpers.ts
